Update prototype functions to use opal namespace
the opal:success, opal:message, and opal:failure functions are for the
UI do display information in colored text to the user. It may feel more
natural to refer to 'failure' as 'error', but there is already a
logging function named error. I want one function to do one thing. So,
to prevent the need for a sub namespace, I use different names.