Create about_popos function
Informs the user about the high-level system information. It checks for
neofetch, so if the user has already installed it, it wll use neofetch.
Otherwise it makes use of it already installed. Becuase these fallback
functions vary in their output, it's impossible to guarantee uniformity.
After all, that's what neofetch is doing.

Issue: #18