Create a template for new commands
Provides some common scaffolding to create a command script, to easily
enable the use of options and arguments. The `help` and `version`
options are the only ones provided with default functionality. The user
is expected to create a copy of this script; then update their copy with
whatever functionaliy they need.

issue #27