Add Chicago Manual of Style (CMOS) formats
Creates a set of formats inspired by The Chicago Manual of Style, 17 ed,
while staying consistent with the Opal format convention of '-date',
'-datetime', and '-timestamp'. The 17th Ed makes suggestions, but
doesn't have a specific rule on formatting the time immediately
following a date, or specifying when to use a timestamp.