Add opal:is_empty and update opal:str_trim
The opal:str_trim will now remove tabs as well as spaces.
Add a new function opal:is_empty which will return true if the value
is null, empty string, or contains only white space.

A value that contains only whitespace is useless, so it should be easy
to check for it. The logic of str_trim is duplicated in is_empty because
an Opal core function cannot have a dependency on a non-core function.